Complete refactor to avoid loading private keys
refs: https://github.com/TryGhost/Toolbox/issues/434

This is a hard fork - we had to completely remove the code which handles OCSP, the TLS challenge, and true certificate renewal.

The internal represnetation of certificates changed from a tls certificate (which has a keypair) to the x509 leaf - with just enough information to know whether the certificate needs to be renewed.